+++ demo/demo_script_08_log_to_file.py
@@ -5,8 +5,10 @@ test file for testing writing the log messages to a file
 import logging
 try:
     from src.tmc_driver.tmc_2209 import *
+    from src.tmc_driver.com._tmc_com_uart import *
 except ModuleNotFoundError:
     from tmc_driver.tmc_2209 import *
+    from tmc_driver.com._tmc_com_uart import *
 
 
 print("---")
@@ -31,13 +33,13 @@ logformatter = logging.Formatter('%(name)s %(asctime)s - %(levelname)s - %(messa
 # initiate the Tmc2209 class
 # use your pins for pin_en, pin_step, pin_dir here
 #-----------------------------------------------------------------------
-if BOARD == Board.RASPBERRY_PI:
+if tmc_gpio.BOARD == Board.RASPBERRY_PI:
     tmc = Tmc2209(TmcEnableControlPin(21), TmcMotionControlStepDir(16, 20), TmcComUart("/dev/serial0"),
                    loglevel=loglevel, log_handlers=[logging_handler], log_formatter=logformatter)
-elif BOARD == Board.RASPBERRY_PI5:
+elif tmc_gpio.BOARD == Board.RASPBERRY_PI5:
     tmc = Tmc2209(TmcEnableControlPin(21), TmcMotionControlStepDir(16, 20), TmcComUart("/dev/ttyAMA0"),
                    loglevel=loglevel, log_handlers=[logging_handler], log_formatter=logformatter)
-elif BOARD == Board.NVIDIA_JETSON:
+elif tmc_gpio.BOARD == Board.NVIDIA_JETSON:
     tmc = Tmc2209(TmcEnableControlPin(13), TmcMotionControlStepDir(6, 5), TmcComUart("/dev/ttyTHS1"),
                    loglevel=loglevel, log_handlers=[logging_handler], log_formatter=logformatter)
 else: